callback graph tests
